remove memory.rs
