+++ src/any.rs
@@ -353,13 +353,13 @@ mod tests {
     use super::*;
     use crate::{
         cache::{CachingStrategy, QUERY_CACHE_TABLE, TABLE_CACHE_TABLE},
-        db_kind::DbType,
-        db_row,
-        db_value::{ColumnMap, DbRow, DbValue, JsonValue, StringRow},
-        memory::{
+        cache::{
             clear_memory_query_cache, clear_memory_table_cache, clear_meta_cache,
             get_memory_query_cache_contents, get_memory_table_cache_contents,
         },
+        db_kind::DbType,
+        db_row,
+        db_value::{ColumnMap, DbRow, DbValue, JsonValue, StringRow},
         params,
     };
     use rand::{
