+++ src/lib.rs
@@ -5,7 +5,6 @@ pub mod cache;
 pub mod core;
 pub mod db_kind;
 pub mod db_value;
-pub mod memory;
 pub mod parse;
 pub mod serde;
 pub mod shared;
